make line

+++ src/Ship.java
@@ -1,6 +1,6 @@
 public class Ship {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Ship(int x, int y) {
         this.x = x;
@@ -23,19 +23,19 @@ public class Ship {
         y++;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
